requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.26
pandas
matplotlib
pillow
tqdm
scikit-learn
scipy
requests
PyYAML
joblib
kagglehub
kagglesdk
fastapi
pydantic
uvicorn[standard]